ocr_qbank/requirements.txt
fastapi==0.115.12
uvicorn== 0.34.0
python-dotenv==1.0.1
langchain==0.3.25
langchain_openai==0.3.18
langchain_google_genai==2.1.4
langchain-groq==0.2.1
langchain-anthropic==0.3.1
langchain_community==0.3.24
langchain-astradb==0.6.0
langgraph==0.4.5
requests==2.32.3
pymupdf==1.25.1
scikit-learn==1.6.0
openpyxl==3.1.5
streamlit==1.41.1
SQLAlchemy==2.0.36
mysqlclient==2.2.6
pymysql==1.1.1
aiomysql==0.2.0
bs4==0.0.2
playwright==1.49.1
w3lib==2.2.1
extruct==0.18.0
redis==5.2.1
boto3==1.38.34
mistralai==1.8.2
PyPDF2==3.0.1